requirements.txt
streamlit
panda
sqlalchemy
plotly
dotenv
mysql-connector-python
matplotlib